requirements.txt
mlflow>=2.3.0
streamlit>=1.22.0
torch>=1.13.0
torchvision>=0.14.0
Pillow>=9.3.0
scikit-learn>=1.1.3
pandas>=1.5.0
numpy>=1.23.0
prefect>=2.0
